speech nagivation between home/favorites functional
