+++ frontend/src/pages/Home/Home.js
@@ -1,5 +1,6 @@
 import React from "react";
 import TodoList from "../../components/TodoList";
+import Speech from "./Speech";
 
 
 
@@ -13,6 +14,7 @@ function Welcome(useSemiPersistentState) {
         <>
             <h1 className='welcome'>Welcome {msg}!</h1>
             <h2 className='question'>What would you like to get done today?</h2>
+            <Speech />
             <TodoList useSemiPersistentState={useSemiPersistentState} />
         </>
     );
