+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-speech-recognition": "^3.10.0"
+  }
+}
